Create a `data_transformers` package and move relevant files there
